[suggestion-mode] Remove the option and always suggest
_is_c_extension already checks if owner is a Module

See https://github.com/pylint-dev/pylint/pull/9962#discussion_r1782040559

Caching start to make sense in _similar_name

See https://github.com/pylint-dev/pylint/pull/9962#discussion_r1782042663

We had to create a clean_lru_cache file to prevent a circular import.
