+++ pylint/utils/utils.py
@@ -40,7 +40,6 @@ DEFAULT_LINE_LENGTH = 79
 
 # These are types used to overload get_global_option() and refer to the options type
 GLOBAL_OPTION_BOOL = Literal[
-    "suggestion-mode",
     "analyse-fallback-blocks",
     "allow-global-unused-variables",
     "prefer-stubs",
